Fix warnings; fix parsing of declination sign
When declination is negative, but less than a degree, say -00 01,
the correct sign for the declination must be chosen; so now the
sign is read separately from the rest of the declination data
from the observation record. Also, warnings are less verbose:
when catalog not found corresponds to reference catalog, no
warning is thrown.